+++ CHANGELOG.md
@@ -4,3 +4,5 @@
 ### Added
 - Basic model.
 - Sync of data with hydrus network.
+- Basic gui, that allow rate image by elo.
+- Generate collage button.